// types/errors.ts
// @/types/errors

export enum PocketBaseAuthErrorType {
    UNKNOWNERROR = "UNKNOWNERROR",
    AUTHORIZATIONCODEMISSING = "CODEMISSING",
    COOKIEMISSING = "COOKIEMISSING",
    TOKENMSSING = "TOKENMSSING",
    POCKETBASEMISSING = "POCKETBASEMISSING",
    USERTOKENMISSING = "USERTOKENMISSING",
    USERRECORDMISSING = "USERRECORDMISSING",
    USERIDMISSING = "USERIDMISSING",
    EMAILMISSING = "EMAILMISSING",
  }